Initial Android Project Commit
This is the intial project structure and main activity view commit
for the instrument maintenance app for android. This implements a
main view with a simple button, dropdown of some predefined strings
that are placholders for the user ids, a "Toast" (new to me), and a
select button. No theming or connecting between any data or anything
has been implemented.